requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
cloudinary==1.22.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1.2
django-bootstrap4==1.0.1
django-braces==1.14.0
django-countries==5.5
django-crispy-forms==1.8.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-registration==2.4.1
django-star-ratings==0.9.1
django-tinymce==2.8.0
djangorestframework==3.12.1
gunicorn==20.0.4
psycopg2==2.8.6
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
swapper==1.1.2.post1
urllib3==1.25.11
whitenoise==5.2.0
